# docs/contracts/poundcake-openapi.json
{
  "components": {
    "schemas": {
...
      "ExecuteRequest": {
        "properties": {
          "context": {
            "additionalProperties": true,
            "title": "Context",
            "type": "object"
          },
          "execution_engine": {
            "maxLength": 50,
            "title": "Execution Engine",
            "type": "string"
          },
          "execution_parameters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Parameters"
          },
          "execution_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Payload"
          },
          "execution_target": {
            "maxLength": 255,
            "title": "Execution Target",
            "type": "string"
          },
          "retry_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Retry Count",
            "type": "integer"
          },
          "retry_delay": {
            "default": 0,
            "minimum": 0.0,
            "title": "Retry Delay",
            "type": "integer"
          },
          "timeout_duration_sec": {
            "default": 300,
            "exclusiveMinimum": 0.0,
            "title": "Timeout Duration Sec",
            "type": "integer"
          }
        },
        "required": [
          "execution_engine",
          "execution_target"
        ],
        "title": "ExecuteRequest",
        "type": "object"
      },
      "ExecutionEnvelopeResponse": {
        "properties": {
          "attempts": {
            "default": 1,
            "title": "Attempts",
            "type": "integer"
          },
          "engine": {
            "title": "Engine",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "execution_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Ref"
          },
          "raw": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "canceled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "engine",
          "status"
        ],
        "title": "ExecutionEnvelopeResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "PoundCake API",
    "version": "2.0.131"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cook/execute": {
      "post": {
        "description": "Generic execution endpoint for all supported execution engines.",
        "operationId": "execute_ingredient_api_v1_cook_execute_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "title": "X-Request-Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionEnvelopeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Ingredient",
        "tags": [
          "infrastructure"
        ]
      }
    },
    "/api/v1/cook/executions": {
      "get": {
        "description": "List StackStorm executions (optional parent filter).",
        "operationId": "list_st2_executions_api_v1_cook_executions_get",
        "parameters": [
          {
            "description": "Filter by parent execution id",
            "in": "query",
            "name": "parent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by parent execution id",
              "title": "Parent"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StackStormExecutionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List St2 Executions",
        "tags": [
          "infrastructure"
        ]
      }
    },
    "/api/v1/cook/executions/{execution_id}": {
      "delete": {
        "description": "Delete a StackStorm execution record.",
        "operationId": "delete_st2_execution_api_v1_cook_executions__execution_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteExecutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete St2 Execution",
        "tags": [
          "infrastructure"
        ]
      },
      "get": {
        "description": "Get a StackStorm execution by ID.",
        "operationId": "get_st2_execution_api_v1_cook_executions__execution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StackStormExecutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get St2 Execution",
        "tags": [
          "infrastructure"
        ]
      },
      "put": {
        "description": "Cancel a StackStorm execution.",
        "operationId": "cancel_st2_execution_api_v1_cook_executions__execution_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteExecutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel St2 Execution",
        "tags": [
          "infrastructure"
        ]
      }
    },
    "/api/v1/cook/executions/{execution_id}/tasks": {
      "get": {
        "description": "Get StackStorm execution task results by ID (Orquesta).",
        "operationId": "get_st2_execution_tasks_api_v1_cook_executions__execution_id__tasks_get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StackStormTaskListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get St2 Execution Tasks",
        "tags": [
          "infrastructure"
        ]
      }
    },
    "/api/v1/cook/packs": {
      "get": {
        "description": "Return the current generated PoundCake StackStorm pack as tar.gz.",
        "operationId": "get_stackstorm_pack_tgz_api_v1_cook_packs_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Pack-Sync-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Pack-Sync-Token"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Stackstorm Pack Tgz",
        "tags": [
          "infrastructure"
        ]
      }
    },
    "/api/v1/cook/sync": {
      "post": {
        "description": "Sync StackStorm actions/workflows to PoundCake Ingredients/Recipes.",
        "operationId": "sync_stackstorm_to_poundcake_api_v1_cook_sync_post",
        "parameters": [
          {
            "description": "Mark bootstrap completion",
            "in": "query",
            "name": "mark_bootstrap",
            "required": false,
            "schema": {
              "default": false,
              "description": "Mark bootstrap completion",
              "title": "Mark Bootstrap",
              "type": "boolean"
            }
          },
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStackStormResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Stackstorm To Poundcake",
        "tags": [
          "infrastructure"
        ]
      }
    },
    "/api/v1/cook/workflows/register": {
      "post": {
        "description": "Register an Orquesta workflow in StackStorm.",
        "operationId": "register_st2_workflow_api_v1_cook_workflows_register_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRegistrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register St2 Workflow",
        "tags": [
          "infrastructure"
        ]
      }
    },
...
  }
}